+++ sheriff/CHANGELOG.md
@@ -1,5 +1,24 @@
 # Revision history for sheriff
 
+## 0.2.1.0
+* Add infinite recursion rule
+* Add test suite for infinite recursion rule detection
+* Add ignore functions in function rules
+* Refactor to use implicit params for passing plugin opts
+* Refactor CommonTypes, Types, Utils and TypeUtils
+* Update documentation for infinite recursion rule and supported patterns
+
+## 0.2.0.6
+* Refactor test cases in cabal file to use smaller test suites
+* Revert functionality to provide sheriff plugin opt in single module
+* Refactor plugin to split types modules
+
+## 0.2.0.5
+* Refactor test cases to different sub test cases instead of single test file
+* Add functionality to provide sheriff plugin opt in single module
+* Refactor & Re-enable log rules check as a subtest
+* Add -fkeep-going to keep testing individual sub-test despite compilation error
+
 ## 0.2.0.4
 * Fix modules names not getting matched due to wildcard support logical bug